Move SAS method related functions to another file
